perf experiment: disable large-move-check
